refactor(home): move page-level sections into reusable components
